Replace uses of jax.util.safe_zip with zip(..., strict=True).
Python 3.10 added a strict=... keyword argument to the zip builtin. See: https://peps.python.org/pep-0618/

This subsumes jax.util.safe_zip, which is equivalent to:

def safe_zip(*args): return list(zip(*args, strict=True))

Provided this code is used under Python 3.10+ only, we can replace uses of jax.util.safe_zip with calls to zip(..., strict=True). Strictly speaking this is not exactly equivalent since zip(...) returns an iterator instead of a list, but that does not appear to matter in the code updated by this change.

This change is in preparation for removing safe_zip from the JAX API.

PiperOrigin-RevId: 507868975