PR #26348: [better_errors] Continue adding debug info to Jaxprs (step 3)
Imported from GitHub PR https://github.com/jax-ml/jax/pull/26348

This follows after #26078, and #26313, adding `debug_info` to more calls to `lu.wrap_init`.

As part of this I have changed the primitives `custom_vjp_call_jaxpr` and `custom_lin` to take the `bwd` parameter as a `lu.WrappedFun`, which carries debug info. Previously, this was a `Callable`, but in almost all cases if was really ` lu.WrappedFun.call_wrapped`.

These changes ensure that all the `lu.wrap_init` and `Jaxpr` are called with debug_info in the `api_test.py:CustomVJPTest` and `api_test.py:CustomJVPTest`.
Copybara import of the project:

--
904b74860ce74723ec2c04a1ecf1b2112327ad3d by George Necula <gcnecula@gmail.com>:

[better_errors] Continue adding debug info to Jaxprs (step 3)

This follows after #26078, and #26313, adding `debug_info` to
more calls to `lu.wrap_init`.

As part of this I have changed the primitives `custom_vjp_call_jaxpr`
and `custom_lin` to take the `bwd` parameter as a `lu.WrappedFun`,
which carries debug info. Previously, this was a `Callable`, but in
almost all cases if was really ` lu.WrappedFun.call_wrapped`.

Merging this change closes #26348

PiperOrigin-RevId: 723920031